Implemented King and Queen logic
Added the queen piece's movement logic which was trivial with some simple changes to the Rook and Bishop functions, then an or check is performed on those for the Queen movement.

Added the king piece's movement logic with castling.

Added the pawn's promotion logic and fixed a slight mistake in the en passant's code.