Implemented Check, Mating and Game Loop logic
This ended up turning into quite a mess and a half because of the "anyMovesLeft" function and others with similar functionality since testing possible states requires a copy of the current board, which can be a little tricky because of references and such. A lot of the other functions were also coded using the Board class' own "cells" attribute which made untangling things messier than they should have.

I need to rewrite and refactor a lot of things before going forward and implementing a GUI which is next on the to do list.