Added placeholder GUI and fixed some bugs
Added a placeholder GUI with Java Swing, it currently just prints the piece labels, I will add some sprites later.

Added a "sendMove" and "getCurrentBoardState" functions to the "Board" class to handle sending and receiving data from the mouse clicks in the GUI. The old terminal method is still there commented if one wants to use it.

I still need to add some way to handle the promotion logic in the GUI.

Bug fixes to the "checkPossibleMovements" function which was causing unintended side effects.

I also need to rewrite a lot of the checking functions to remove the necessity for the janky, "copy" ones.