Some refactoring and bug fixes
Broadly changed most functions to take in a reference to the 2D array of "Cell" as an argument, this allows for smoother operations when working with copied states, and also removes the necessity for the janky "copy" functions.

Changed the "turn" boolean into an enum, more clearly indicating "currentTurn.WHITE" or "currentTurn.BLACK".

Removed the "possibleCastle" function since you can't actually castle if you are in check.

Rewrote the "isValidWhitePawnMove" and "isValidBlackPawnMove" functions into a single "isValidPawnMove" function.

Rewrote the "isValidRookMove" and "isValidKnightMove" into more concise versions with the same functionality, will look into doing the same for "isValidBishopMove" later.

Fixed a bug in "isValidKingMove" so that now it's able to handle capturing pieces properly. Before it would always be able to capture, now it takes into account whether that space will put them into check or not, this was already done for empty squares.

Fixed a bug in "moveGetsOutofCheck" where it was always considering the previous king's position even if a user had moved it.

And some other minor changes.