Finished small refactoring attempt
Moved piece specific methods to their own classes.

This required creating a new file for the Enums and a number of getter methods for some Board class parameters.

Not really sure if this is really much better or worse, but now I should focus on adding more functionality to the game.