Added a very simple AI opponent
Currently added a very crude AI opponent implementation which will essentially just choose a random valid move available to it, I'll look into improving it with evaluation functions in the future.

I also need to add a few configuration options such as human or AI opponent in the main menu and whether the AI will play with the white or black pieces.